requirements.txt
Numpy
Pytorch
Pytorch-Geometric
